requirements.txt
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
django==4.2.6 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.1.0 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
uWSGI==2.0.21 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.6
python-dotenv==0.20.0